+++ src/main/java/com/se2023/backend/mapper/UserMapper.java
@@ -8,7 +8,9 @@ import java.util.List;
 @Mapper
 public interface UserMapper {
     @Update("update user set balance = #{balance} where id = #{id}")
-    void updateBalance(Integer id, Integer balance);
+    void updateBalance(Integer id, Double balance);
+
+
 
     @Select("select * from user where username = #{username}")
     List<User> queryUserByName(String username);
